Tarification_service isoler/requirements.txt
transformers 
torch 
torchvision
opencv-python 
pandas 
rapidfuzz
pandas 
fuzzywuzzy 
python-Levenshtein
easyocr
fastapi
uvicorn
matplotlib
pillow
python-multipart
statsmodels
scikit-learn